+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Refined Storage Changelog
 
 ### 1.4.5
+- Updated Forge to 2296 (raoulvdberge)
 - Fixed Crafting Tweaks buttons positioned wrongly (blay09)
 - Fixed Crafting Tweaks keybindings interfering with RS keybindings (blay09)
 - Fixed crash when updating storages (raoulvdberge)