+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Fixed crash when updating storages (raoulvdberge)
 - Fixed no tooltips for fluid filter slots (raoulvdberge)
 - Fixed Disk Manipulator in fluid mode not showing fluids (raoulvdberge)
+- Fixed dupe bug in the Interface (raoulvdberge)
 - Removed ticking tile entities, every tile entity in RS is non-ticking now (raoulvdberge)
 
 ### 1.4.4